[KAN-33] fix: mock data 파일로 관리 및 드롭다운이 열렸을 때 Filter 를 다시 클릭하면 안 닫히는 버그 수정
