+++ src/app/(main)/gatherings/page.tsx
@@ -10,38 +10,7 @@ import FilterList from '@/app/components/Filter/FilterList';
 import FilterSort from '@/app/components/Filter/FilterSort';
 import CardList from '@/app/components/CardList/CardList';
 
-const tomorrow = new Date();
-tomorrow.setDate(tomorrow.getDate() + 1);
-
-const registrationEnd = new Date(tomorrow);
-registrationEnd.setDate(tomorrow.getDate() - 1);
-
-const mockData = {
-  teamId: 101,
-  id: 1,
-  type: 'Fitness',
-  name: 'Fitness Gathering',
-  dateTime: tomorrow.toISOString(),
-  registrationEnd: registrationEnd.toISOString(),
-  location: 'Seoul, Korea',
-  participantCount: 8,
-  capacity: 15,
-  image: '/images/mock-image.png',
-  createdBy: 1001,
-  canceledAt: registrationEnd.toISOString(),
-};
-
-const OPTIONS = [
-  '건대입구',
-  '을지로 3가',
-  '신림',
-  '홍대입구',
-  '시청',
-  '신대방',
-  '서울대입구',
-];
-
-const SORT_OPTIONS = ['마감 임박', '을지로 3가', '참여 인원 순'];
+import { mockData, OPTIONS, SORT_OPTIONS } from './mockData/mockData';
 
 const GatheringsPage = () => {
   const [activeTab, setActiveTab] = useState<'workation' | 'dalaemfit'>(
@@ -104,6 +73,7 @@ const GatheringsPage = () => {
               <FilterList state='default' options={OPTIONS}>
                 지역 전체
               </FilterList>
+              {/* TODO: 클릭 시 date picker 추가 */}
               <FilterList state='default'>날짜 전체</FilterList>
             </div>
             <div>
