+++ src/app/components/DropDown/DropDown.tsx
@@ -1,7 +1,5 @@
 'use client';
 
-import { useRef, useEffect } from 'react';
-
 interface DropDownProps {
   options: string[];
   onSelect: (option: string) => void;
@@ -9,34 +7,13 @@ interface DropDownProps {
 }
 
 const DropDown = ({ options, onSelect, onClose }: DropDownProps) => {
-  const dropDownRef = useRef<HTMLDivElement>(null);
-
   const handleSelect = (option: string) => {
     onSelect(option);
     onClose();
   };
 
-  useEffect(() => {
-    const handleClickOutside = (event: MouseEvent) => {
-      if (
-        dropDownRef.current &&
-        !dropDownRef.current.contains(event.target as Node)
-      ) {
-        onClose();
-      }
-    };
-
-    document.addEventListener('mousedown', handleClickOutside);
-    return () => {
-      document.removeEventListener('mousedown', handleClickOutside);
-    };
-  }, [onClose]);
-
   return (
-    <div
-      className='absolute z-10 mt-4 max-h-240 w-full overflow-y-auto rounded-xl bg-var-gray-50 ring-2 ring-var-gray-400'
-      ref={dropDownRef}
-    >
+    <div className='absolute z-10 mt-4 max-h-240 w-full overflow-y-auto rounded-xl bg-var-gray-50 ring-2 ring-var-gray-400'>
       <ul>
         {options.map((option, index) => (
           <li
