+++ src/app/components/Filter/FilterSort.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { useState } from 'react';
+import { useRef, useState, useEffect } from 'react';
 
 import { IconSort } from '@/public/icons';
 import DropDown from '../DropDown/DropDown';
@@ -25,6 +25,8 @@ const FilterSort = ({
   const [selectedOption, setSelectedOption] = useState<string | null>(null);
   const [currentState, setCurrentState] = useState<'default' | 'active'>(state);
 
+  const containerRef = useRef<HTMLDivElement>(null);
+
   const handleOptionSelect = (option: string) => {
     setSelectedOption(option);
     setIsOpen(false);
@@ -35,10 +37,27 @@ const FilterSort = ({
     setIsOpen((prev) => !prev);
   };
 
+  // 외부 클릭 시 드롭다운 닫힘
+  useEffect(() => {
+    const handleClickOutside = (event: MouseEvent) => {
+      if (
+        containerRef.current &&
+        !containerRef.current.contains(event.target as Node)
+      ) {
+        setIsOpen(false);
+      }
+    };
+
+    document.addEventListener('mousedown', handleClickOutside);
+    return () => {
+      document.removeEventListener('mousedown', handleClickOutside);
+    };
+  }, []);
+
   return (
-    <div className='relative'>
+    <div className='relative' ref={containerRef}>
       <div
-        className={`flex h-36 w-36 items-center justify-between rounded-[12px] p-[6px] text-[13px] font-medium md:h-40 md:w-120 md:px-12 md:py-[6px] ${stateClasses[currentState]}`}
+        className={`flex h-36 w-36 cursor-pointer items-center justify-between rounded-[12px] p-[6px] text-[13px] font-medium md:h-40 md:w-120 md:px-12 md:py-[6px] ${stateClasses[currentState]}`}
         onClick={toggleDropDown}
       >
         <IconSort className='h-24 w-24' />
